geolocation/geo-00059.html

<!DOCTYPE html>
<html>
  <head>
    <title>Geolocation Test: call watchPosition() with wrong type for first argument, exception expected</title>
    <link rel="author" title="Intel" href="http://www.intel.com/">
    <link rel="help" href="http://www.w3.org/TR/2012/PR-geolocation-API-20120510/#geolocation">
    <link rel="help" href="http://www.w3.org/TR/2012/PR-geolocation-API-20120510/#watch-position">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script>
      var t =async_test("Test checks that call watchPosition() with wrong type for first argument, exception expected");
      setup({timeout: 20000});
      try {
        window.navigator.geolocation.watchPosition(3);
      } catch (err) {
        t.step(function () {
          assert_true(true, "call watchPosition() with wrong type for first argument, exception expected");
        });
        t.done();
      }
    </script>
  </body>
</html>
